move token_based_document_with_entities_and_relations_to_text_based back to src.document.conversion
